Refactor file name and structure to match scenarios. Leaving other tests in place that could be later removed/updated to match new structure
